Working on main class

+++ GeneticAlgorithm/src/InstanceFileReader.java
@@ -7,21 +7,21 @@ import org.jorlib.io.tspLibReader.graph.DistanceTable;
 
 public class InstanceFileReader {
 
-	public double[][] read(String fileName) throws IOException {
+	public Integer[][] read(String fileName) throws IOException {
 		File instnaceFile = new File(fileName);
 		TSPLibInstance readInstance = new TSPLibInstance(instnaceFile);
 
 		return convert(readInstance.getDistanceTable(), fileName);
 	}
 
-	private double[][] convert(DistanceTable distanceTable, String instanceName) {
+	private Integer[][] convert(DistanceTable distanceTable, String instanceName) {
 		int instanceSize = distanceTable.listNodes().length;
-		double[][] matrix = new double [instanceSize][instanceSize];
+		Integer[][] matrix = new Integer [instanceSize][instanceSize];
 		
 		for (int i = 0; i < instanceSize; i++) {
 			for (int j = 0; j < instanceSize; j++) {
 				
-				matrix[i][j] = distanceTable.getDistanceBetween(i, j);
+				matrix[i][j] = (int) distanceTable.getDistanceBetween(i, j);
 			}
 		}
